Refactor for orientation
Now, orientation is provided in grid slot. It allows to use the same orientable body_data (e.g. comet) and have different orientations.

This method has a big issue because i had to change the signature of all function constraints, even for body that don't need orientation.
I didn't come up with a better idea for now, so I implemented it