Temporary ugly fix to visual bug
When a drag preview is created in a container, it appears first at the side left of the screen before being at the right position.

It seems to be a bug of godot. I didn't find a good fix for now, so I'm using a very ugly fix I found on the internet to solve it for now